+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQHF19CQFVDK608DCW33G87Z"
+        "01KQM5RB3DREMXCME0R9EXJBBC"
     )
     r"""The unique identifier of the data source"""