+++ packages/orq-rc/RELEASES.md
@@ -7499,3 +7499,13 @@ Based on:
 - [python v4.9.0-rc.17] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.17] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.17 - packages/orq-rc
+
+## 2026-05-02 11:06:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.11 (2.881.17) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.18] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.18] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.18 - packages/orq-rc